+++ ghost/core/core/server/services/mail-events/libraries.d.ts
