+++ ghost/core/package.json
@@ -97,6 +97,7 @@
     "@tryghost/identity-token-service": "0.0.0",
     "@tryghost/image-transform": "1.3.0",
     "@tryghost/importer-handler-content-files": "0.0.0",
+    "@tryghost/in-memory-repository": "0.0.0",
     "@tryghost/job-manager": "0.0.0",
     "@tryghost/kg-card-factory": "5.1.0",
     "@tryghost/kg-converters": "1.1.0",
@@ -111,7 +112,6 @@
     "@tryghost/link-replacer": "0.0.0",
     "@tryghost/logging": "2.4.21",
     "@tryghost/magic-link": "0.0.0",
-    "@tryghost/mail-events": "0.0.0",
     "@tryghost/mailgun-client": "0.0.0",
     "@tryghost/member-attribution": "0.0.0",
     "@tryghost/member-events": "0.0.0",
