--- ghost/mail-events/README.md
@@ -1,17 +1,0 @@
-# Mail Events
-
-Processes mail events
-
-## Develop
-
-This is a monorepo package.
-
-Follow the instructions for the top-level repo.
-
-1. `git clone` this repo & `cd` into it as usual
-2. Run `yarn` to install top-level dependencies
-
-## Test
-
-- `yarn lint` run just eslint
-- `yarn test` run lint and tests
