--- ghost/mail-events/package.json
@@ -1,41 +1,0 @@
-{
-    "name": "@tryghost/mail-events",
-    "version": "0.0.0",
-    "repository": "https://github.com/TryGhost/Ghost/tree/main/packages/mail-events",
-    "author": "Ghost Foundation",
-    "private": true,
-    "main": "build/index.js",
-    "types": "build/index.d.ts",
-    "scripts": {
-        "build": "yarn build:ts",
-        "build:ts": "tsc",
-        "test:unit": "NODE_ENV=testing c8 --src src --all --check-coverage --100 --reporter text --reporter cobertura -- mocha --reporter dot -r ts-node/register './test/**/*.test.ts'",
-        "test": "yarn test:types && yarn test:unit",
-        "test:types": "tsc --noEmit",
-        "lint:code": "eslint src/ --ext .ts --cache",
-        "lint": "yarn lint:code && yarn lint:test",
-        "lint:test": "eslint -c test/.eslintrc.js test/ --ext .ts --cache"
-    },
-    "files": [
-        "build"
-    ],
-    "devDependencies": {
-        "c8": "8.0.1",
-        "mocha": "10.2.0",
-        "sinon": "15.2.0"
-    },
-    "dependencies": {
-        "@tryghost/errors": "1.3.5",
-        "@tryghost/in-memory-repository": "0.0.0",
-        "@tryghost/tpl": "0.1.32"
-    },
-    "c8": {
-        "exclude": [
-            "src/index.ts",
-            "src/MailEventRepository.ts",
-            "src/InMemoryMailEventRepository.ts",
-            "src/**/*.d.ts",
-            "test/**/*.ts"
-        ]
-    }
-}
