--- ghost/mail-events/src/index.ts
@@ -1,3 +1,0 @@
-export * from './MailEvent';
-export * from './MailEventRepository';
-export * from './MailEventService';
